Moved all back end stuff to backend package
