+++ src/main/java/com/group17/hifiprototype/backend/dataclasses/Race.java
@@ -1,20 +1,16 @@
-package com.group17.hifiprototype.dataclasses;
+package com.group17.hifiprototype.backend.dataclasses;
 
-import com.group17.hifiprototype.weatherapi.WeatherLoader;
-import com.group17.hifiprototype.weatherapi.weather;
+import com.group17.hifiprototype.backend.weatherapi.WeatherLoader;
+import com.group17.hifiprototype.backend.weatherapi.weather;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.apache.commons.lang3.StringUtils;
 
 import java.io.IOException;
 import java.time.Duration;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.*;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
