+++ src/main/java/com/group17/hifiprototype/backend/dataclasses/RaceLoader.java
@@ -1,11 +1,9 @@
-package com.group17.hifiprototype.dataclasses;
+package com.group17.hifiprototype.backend.dataclasses;
 
 import org.json.JSONObject;
 
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
-import java.io.Reader;
 import java.nio.file.Files;
 import java.util.*;
 import java.util.stream.Collectors;
